requirements.txt
flask_cors==6.0.1
kagglehub==0.3.13
matplotlib==3.10.7
mlflow==3.5.1
mlflow_skinny==3.5.1
mlflow_tracing==3.5.1
numpy==2.3.4
pandas==2.3.3
PyYAML==6.0.3
scikit_learn==1.7.2
torch==2.9.0+cu130
transformers==4.57.1
